requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
babel==2.18.0
bangla==0.0.5
blinker==1.9.0
blis==1.3.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.13
Cython==3.2.4
dateparser==1.1.8
decorator==5.2.1
docopt==0.6.2
einops==0.8.2
encodec==0.1.1
filelock==3.24.3
Flask==3.1.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
g2pkk==0.1.2
grpcio==1.78.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.3
jsonlines==1.2.0
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
murmurhash==1.0.15
networkx==2.8.8
nltk==3.9.2
num2words==0.5.14
numba==0.64.0
numpy==1.26.4
openai-whisper==20250625
packaging==26.0
pandas==1.5.3
pillow==12.1.1
platformdirs==4.9.2
pooch==1.9.0
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.3.2
pypinyin==0.55.0
pysbd==0.3.4
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
SudachiDict-core==20260116
SudachiPy==0.6.10
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.15.2
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.3
trainer==0.0.36
transformers==4.36.2
TTS==0.22.0
typeguard==4.5.1
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
umap-learn==0.5.11
Unidecode==1.4.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
Werkzeug==3.1.6
wrapt==2.1.1
yarl==1.22.0
